split fuzz and export
